requirements.txt
echo "
fastapi
pydantic
ibm-cloud-sdk-core
requests
pillow